# Conflicts:
#	client/src/main/java/client/scenes/AddEventCtrl.java
#	client/src/main/java/client/scenes/AdminOverviewCtrl.java
#	client/src/main/java/client/scenes/EventInfoCtrl.java
#	client/src/main/java/client/scenes/ExpenseInfoCtrl.java
#	client/src/main/java/client/scenes/IntroPageCtrl.java
#	client/src/main/java/client/scenes/UserDebtCtrl.java
#	client/src/main/java/client/services/AdminOverviewService.java
#	client/src/main/resources/CONFIG.json
#	client/src/main/resources/locales/resource_de.properties
#	client/src/main/resources/locales/resource_en.properties
#	client/src/main/resources/locales/resource_es.properties
#	client/src/main/resources/locales/resource_fr.properties
#	client/src/main/resources/locales/resource_nl.properties
#	client/src/main/resources/locales/resource_tr.properties

+++ client/src/main/java/client/scenes/IntroPageCtrl.java
@@ -8,6 +8,7 @@ import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
 import javafx.stage.Stage;
 import javafx.scene.control.Button;
 import javafx.scene.control.TextField;
@@ -15,6 +16,7 @@ import javafx.scene.input.KeyEvent;
 import javafx.util.Duration;
 
 import java.net.URL;
+import java.util.Optional;
 import java.util.ResourceBundle;
 
 
@@ -55,16 +57,15 @@ public class IntroPageCtrl implements Initializable {
     /**
      * Method to handle the start button
      */
-    public void start(){
+    public void start() {
         ServerUtils server = new ServerUtils();
-        if(!server.setServerAddress(serverAddress.getText())){
+        if (!server.setServerAddress(serverAddress.getText())) {
             Alert alert = new Alert(Alert.AlertType.ERROR);
             alert.setTitle(service.getString("server-not-found"));
             alert.setHeaderText(service.getString("the-server-you-wanted-is-unavailable"));
             alert.setContentText(service.getString("please-check-the-server-address-and-try-again"));
             alert.show();
-        }
-        else{
+        } else {
             Stage stage = (Stage) startButton.getScene().getWindow();
             // do what you have to do
             stage.close();
@@ -74,38 +75,45 @@ public class IntroPageCtrl implements Initializable {
 
     /**
      * Method to process the server address
+     *
      * @param actionEvent the clicking of the button
      */
-    public void processServer(ActionEvent actionEvent){
+    public void processServer(ActionEvent actionEvent) {
         ServerUtils server = new ServerUtils();
         String address = serverAddress.getText();
-        if(server.setServerAddress(address)){
-            serverAddressButton.setText("\u2713");
-            serverAddressButton.setStyle("-fx-background-color:  green");
-            PauseTransition pt = new PauseTransition(Duration.seconds(3.0));
-            pt.setOnFinished(e -> {
-                serverAddressButton.setText("\u2192");
-                serverAddressButton.setStyle("-fx-background-color:  fd7f20");
-                service.getMainCtrl().writeServerAddressToConfigFile(address);
 
-            });
-            pt.play();
-            service.getMainCtrl().deleteAllEventsFromConfig();
-        }
-        else{
-            serverAddressButton.setText("\u274C");
-            serverAddressButton.setStyle("-fx-background-color:  D11A2A");
-            PauseTransition pt = new PauseTransition(Duration.seconds(3.0));
-            pt.setOnFinished(e -> {
-                serverAddressButton.setText("\u2192");
-                serverAddressButton.setStyle("-fx-background-color:  fd7f20");
-            });
-            pt.play();
-            Alert alert = new Alert(Alert.AlertType.ERROR);
-            alert.setTitle(service.getString("server-not-found"));
-            alert.setHeaderText(service.getString("the-server-you-wanted-is-unavailable"));
-            alert.setContentText(service.getString("please-check-the-server-address-and-try-again"));
-            alert.show();
+        Alert conf = new Alert(Alert.AlertType.CONFIRMATION);
+        conf.setTitle("Confirmation Dialog");
+        conf.setContentText("Are you sure you want to connect to the server : " + address + " ?");
+        Optional<ButtonType> result = conf.showAndWait();
+        if (result.get() == ButtonType.OK) {
+            if (server.setServerAddress(address)) {
+                serverAddressButton.setText("\u2713");
+                serverAddressButton.setStyle("-fx-background-color:  green");
+                PauseTransition pt = new PauseTransition(Duration.seconds(3.0));
+                pt.setOnFinished(e -> {
+                    serverAddressButton.setText("\u2192");
+                    serverAddressButton.setStyle("-fx-background-color:  fd7f20");
+                    service.getMainCtrl().writeServerAddressToConfigFile(address);
+
+                });
+                pt.play();
+                service.getMainCtrl().deleteAllEventsFromConfig();
+            } else {
+                serverAddressButton.setText("\u274C");
+                serverAddressButton.setStyle("-fx-background-color:  D11A2A");
+                PauseTransition pt = new PauseTransition(Duration.seconds(3.0));
+                pt.setOnFinished(e -> {
+                    serverAddressButton.setText("\u2192");
+                    serverAddressButton.setStyle("-fx-background-color:  fd7f20");
+                });
+                pt.play();
+                Alert alert = new Alert(Alert.AlertType.ERROR);
+                alert.setTitle(service.getString("server-not-found"));
+                alert.setHeaderText(service.getString("the-server-you-wanted-is-unavailable"));
+                alert.setContentText(service.getString("please-check-the-server-address-and-try-again"));
+                alert.show();
+            }
         }
     }
 
@@ -117,7 +125,7 @@ public class IntroPageCtrl implements Initializable {
     public void keyPressed(KeyEvent e) {
         switch (e.getCode()) {
             case ENTER:
-                if(serverAddress.isFocused()){
+                if (serverAddress.isFocused()) {
                     processServer(null);
                 }
             case ESCAPE:
